mock-servicio/bff_requirements.txt
aiopulsar-py==0.1.0a0
fastapi~=0.110.0
pydantic~=2.6.3
pydantic_settings
uvicorn[standard]==0.20.0
pulsar-client
pulsar-client[avro]
requests==2.28.2
sse-starlette==1.2.1
strawberry-graphql[fastapi]==0.156.3